extend check for null values

+++ src/main/java/de/idsmannheim/lza/datacitejavaapi/DataciteAPITools.java
@@ -50,7 +50,7 @@ public class DataciteAPITools {
         Dois dois = api.ListAllDOIs(prefix);
         LOG.info(dois.toJson());
         for (DoisDataInner data : dois.getData()) {
-            if (data.getAttributes().getState().equals("draft")) {
+            if (data != null && data.getAttributes() != null && data.getAttributes().getState() != null && data.getAttributes().getState().equals("draft")) {
                 LOG.info("Deleting " + data.getId());
                 api.deleteDraftDOI(data.getId());
             }
